% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSmoothSplineLine.r
\name{addSmoothSplineLine}
\alias{addSmoothSplineLine}
\title{Add a fitted cubic smoothing spline to the supplied graph}
\usage{
addSmoothSplineLine(y, x = NULL, col = "blue", lwd = 2, ...)
}
\arguments{
\item{y}{yvalues}

\item{x}{(optional) x values}
}
\description{
Add a fitted cubic smoothing spline to the supplied graph
}
\details{
...
}
\examples{
n <- 1e3
dat <- data.frame(
  x = (1:n)+50,
  y = sin(seq(0, 5*pi, length.out = n)) + rnorm(n=n, mean = 0, sd=0.1)
)
plot(dat)
addSmoothSplineLine(y=dat$y,x=dat$x,col='blue')
}
\author{
JuG
}
