% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bprop2.r
\name{B.prop2}
\alias{B.prop2}
\title{Bayesian comparison for two proportions}
\usage{
B.prop2(
  var1,
  var2,
  n.iter = 11000,
  n.burnin = 1000,
  n.thin = 1,
  prior.beta = c(1, 1, 1, 1)
)
}
\arguments{
\item{var1}{variable 1}

\item{var2}{= variable 2}

\item{n.iter}{number of iteration}

\item{n.burnin}{number of iteration used in the update of the model}

\item{n.thin}{n thin}

\item{prior.beta}{beta priors (vector of 4 parameters, two first for var1)}
}
\value{

}
\description{
Cod to compare two proportions using bayesian model with beta priors
}
\examples{
dtf<-data.frame(grp=c(rep("A",150),rep("B", 150)),
                res=c(rbinom(n = 150, size = 1, prob = .5),
                      rbinom(n = 150, size = 1, prob = .49)) )

res <- B.prop2(var1 = dtf$res[dtf$grp=="A"], var2 = dtf$res[dtf$grp=="B"])
res
}
\author{
JuG
}
