% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B2mean.r
\name{B2mean}
\alias{B2mean}
\title{Bayesian comparison of two means}
\usage{
B2mean(
  var1,
  var2,
  n.iter = 11000,
  n.burnin = 1000,
  n.thin = 1,
  prior.norm = c(0, 0.001, 0, 0.001),
  prior.gamma = c(0.5, 0.5, 0.5, 0.5)
)
}
\arguments{
\item{var1}{variable 1}

\item{var2}{= variable 2}

\item{n.iter}{number of iteration}

\item{n.burnin}{number of iteration used in the update of the model}

\item{n.thin}{n thin}

\item{prior.norm}{normal parameters for priors of means 1 and 2 - two parameters each (mean, precision)}

\item{prior.gamma}{gamma parameters for priors of tau 1 and tau 2 - two parameters each (shape, rate)}
}
\value{
list with Stats (descriptive stats) and MCMC (inference)
}
\description{
Bayesian comparison of two means
}
\details{
precision is the inverse of variance
}
\examples{
x <- rnorm(200, 2.1, .4)
y <- rnorm(200, 2, .4)
B2mean(var1=x, var2=y)

}
\author{
JuG
}
