% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B2prop.r
\name{B2prop}
\alias{B2prop}
\title{Bayesian comparison of two proportions}
\usage{
B2prop(
  var1,
  var2,
  n.iter = 11000,
  n.burnin = 1000,
  n.thin = 1,
  prior.beta = c(1, 1, 1, 1)
)
}
\arguments{
\item{var1}{variable 1}

\item{var2}{= variable 2}

\item{n.iter}{number of iteration}

\item{n.burnin}{number of iteration used in the update of the model}

\item{n.thin}{n thin}

\item{prior.beta}{beta priors (vector of 4 parameters, two first for var1)}
}
\value{
list with Stats (descriptive stats) and MCMC (inference)
}
\description{
Cod to compare two proportions using bayesian model with beta priors
}
\examples{
dtf<-data.frame(grp=c(rep("A",15),rep("B", 15)),
                res=c(rbinom(n = 15, size = 1, prob = .5),
                      rbinom(n = 15, size = 1, prob = .49)) )

res <- B2prop(var1 = dtf$res[dtf$grp=="A"], var2 = dtf$res[dtf$grp=="B"])
res
}
\author{
JuG
}
