% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barycenter_distance.R
\name{barycenter_distance}
\alias{barycenter_distance}
\title{Compute the barycenter distance}
\usage{
barycenter_distance(master, ..., column = dplyr::everything())
}
\arguments{
\item{column}{}
}
\description{
Compute the barycenter distance
}
