% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlandAltman.r
\name{BlandAltman}
\alias{BlandAltman}
\title{Perform and plot Band and Altman analysis}
\usage{
BlandAltman(data1, data2, alpha = 0.05, rep.meas = FALSE, subject,
  silent = FALSE, ...)
}
\arguments{
\item{data1}{vector of data1}

\item{data2}{vector of data2}

\item{alpha}{type I error}

\item{rep.meas}{boolean true if data are repeated measurements}

\item{silent}{boolean if FALSE, parameters list is not returned}
}
\description{
Perform and plot Band and Altman analysis
}
\examples{
data1 <- rnorm(20)
data2 <- rnorm(20,mean = 1)
BlandAltman(data1,data2, ylim=c(-5,5))
}
\author{
JuG
}
