% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_character}
\alias{check_character}
\title{Title Check if character variables}
\usage{
check_character(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
If there are at least one character variables into the data.frame `data`,
this function will raise an error and force to stop the computation.
}
