% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanUpAttempt.R
\name{cleanUpAttempt}
\alias{cleanUpAttempt}
\title{Attempt to clean-up messy vectors}
\usage{
cleanUpAttempt(messy, cluster = FALSE, ngroup = NULL, graph = TRUE)
}
\arguments{
\item{messy}{a factor vector}

\item{cluster}{logical}

\item{ngroup}{integer giving the number of cluster to be formed}

\item{graph}{logical plot or not hclust}
}
\value{
void or cleaned-up factor
}
\description{
An attempt to clean-up messy vectors
}
\details{
use stringdist package. examples from \url{https://cran.r-project.org/web/packages/rrefine/vignettes/rrefine-vignette.html}
}
\examples{
x <- c("Y", "Y,", "Yes", "N", "No",NA,"No","No","No","Nope","Yes","Yes","Yes")
cleanUpAttempt(x)
cleanUpAttempt(x,ngroup =  2)
xc <- cleanUpAttempt(messy = x,ngroup =  2, cluster = TRUE)
summary(xc)


# install.packages("devtools")
devtools::install_github("vpnagraj/rrefine")
library(rrefine)
summary(lateformeeting$what.day.whas.it) #What a mess!
cleanUpAttempt(messy = lateformeeting$what.day.whas.it)
cleaned <-cleanUpAttempt(messy = lateformeeting$what.day.whas.it,ngroup = 5,cluster=TRUE)
summary(cleaned)

cleanUpAttempt(messy = lateformeeting$was.i.on.time.for.work,ngroup=2)

raw <- c("persistante modérée à sévère", "Persistante modérée a sévère","légère", "persistante modérée à sévère",
"persistante légère", "persistante modérée à sévère","persistante  modérée à sévère",
"persistante modérée à sévère","persistante modérée  à sévère", "persistante modérée à sévère","rien à voir",
"persistante modérée à sévère",NA, "persistante modérée à sévère",
"persistante légère","persistante modérée à sévère", "autre solution","persistante modérée,à sévère",
"persistante modérée à sévère","persistante modérée à sévère", "persistante légère","persistante modérée à sévèr",
"persistante légère", "persistant modérée à sévère")
cleanUpAttempt(messy = raw)
cleanUpAttempt(messy = raw,ngroup=4)
}
\author{
JuG
}
