% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeRefine.r
\name{closeRefine}
\alias{closeRefine}
\title{Close connection to openRefine API}
\usage{
closeRefine(dataFrameOutput = TRUE, kill = FALSE)
}
\arguments{
\item{dataFrameOutput}{if TRUE, return a data.frame (a tibble otherwise)}

\item{kill}{if TRUE delete openRefine project}
}
\value{
data.frame or tibble
}
\description{
Get openRefine cleaned data.frame and close
}
\details{
...
}
\examples{
cleanedData <- closeRefine(dataFrameOutput = TRUE)

}
\author{
JuG
}
