% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contColor.r
\name{contColor}
\alias{contColor}
\title{Encode colour for plotting a continuous variabe}
\usage{
contColor(num, palette = matlab.like(100))
}
\arguments{
\item{num}{a continuous variable}

\item{palette}{a palette like terrain.color(n), ...}
}
\description{
Encode colour for plotting a continuous variabe
}
\examples{
x <- runif(n = 100, min = 1, max = 100)
plot(x,x,col= contColor(num = x))
plot(x,x,col= contColor(num = x, palette = terrain.colors(10)))
}
\author{
JuG
}
