% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{create_alpha_index}
\alias{create_alpha_index}
\title{Create names that could be ordered alphabetically}
\usage{
create_alpha_index(n = 10, prefix, sep = "_")
}
\arguments{
\item{n}{An integer. The number of variables}

\item{prefix}{A character. The prefix added before each incremental value.}

\item{sep}{A character. The separator used between the prefix and the numerical
values.}
}
\value{
The function returns a vector of well-defined alphabetically-ordered
names.
}
\description{
Create an ordered list of names using padding of incremental values.
}
