% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{create_kfold}
\alias{create_kfold}
\title{Create k-folds partition of a training data.frame.}
\usage{
create_kfold(X, k = 10, strata, seed)
}
\arguments{
\item{X}{The data.frame to be partitioned.}

\item{k}{Integer. The number of folds}

\item{strata}{Character. The name of the column to stratify the data.frame
with.}

\item{seed}{A number. The random seed.}
}
\value{

}
\description{
Create k-folds partition of a training data.frame.
}
