% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_DMatrix.R
\name{df_to_DMatrix}
\alias{df_to_DMatrix}
\title{Turning data.frame to xgboost::xgb.DMatrix format.}
\usage{
df_to_DMatrix(plsmod, data, additional_columns, threshold = 0.99, y_name)
}
\arguments{
\item{plsmod}{PLS model computed using \code{\link[pls]{mvr}}.}

\item{data}{Data.frame}

\item{additional_columns}{Vector of characters. The list of variable to be
kept from the `data`.}

\item{threshold}{Integer. Minimum cumulative variance that must be explained
by the PLS model components.}

\item{y_name}{Character. Response variable name.}
}
\value{
Matrix of the form \code{\link[xgboost]{xgb.DMatrix}}
}
\description{
Turning data.frame to xgboost::xgb.DMatrix format.
}
