% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawTttSeq.r
\name{drawTttSequence}
\alias{drawTttSequence}
\title{Draw treatment sequence}
\usage{
drawTttSequence(x = seqx, returnType = "init")
}
\arguments{
\item{x}{a treatment sequence (made of 0 and 1)}

\item{returnType}{define which vector to return (init of term)}
}
\value{
vector
}
\description{
Draw treatment sequence for multiple and intermittent exposures (treatment switches) in the estimation of IPW using a survival model. Based on two vectors defining if treatment starts, is continued, is stopped or if no treatment on the time interval.
}
\details{
init takes values of 1 (treatment started), 0 (no treatment) or NA (if term is non missing) ; term takes values of 1 (treatment stopped), 0 (treatment continuation) or NA (if init is non missing)
}
\examples{
seqx <- c(1,1,1,0,0,1,1,0,1,1)
seqx <- c(0, 1, 0, 0, 1, 1, 1, 1)
seqx <- c(0,0,0,0,0)
seqx <- c(1,1,1,1,1)
init <- drawTttSequence(x = seqx, returnType = "init")
term <- drawTttSequence(x = seqx, returnType = "term")
rbind(seqx, init, term)
}
\author{
JuG
}
