% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{factor_to_integer}
\alias{factor_to_integer}
\title{Title Turning factor variables from a data.frame into integer.}
\usage{
factor_to_integer(data)
}
\arguments{
\item{data}{A data.frame in which there are factor to transform.}
}
\value{

}
\description{
Title Turning factor variables from a data.frame into integer.
}
