% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factOrNot.R
\name{factOrNot}
\alias{factOrNot}
\title{Define class of data.frame variables}
\usage{
factOrNot(datafr, arg = "factor", ind = FALSE)
}
\arguments{
\item{datafr}{data.frame}

\item{arg}{class to be evaluated c("factor","Date","integer","numeric","character")}

\item{ind}{if TRUE return index, else return names}
}
\value{
list or names of columns
}
\description{
Merge several data.frames
}
\examples{
library(survival)
data(cgd)
factOrNot(cgd)
factOrNot(cgd,arg = "Date",ind=T)
}
\author{
JuG
}
