% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_validation_set.R
\name{get_validation_set}
\alias{get_validation_set}
\title{Get validation set}
\usage{
get_validation_set(data, variables = dplyr::everything(),
  bound = c(0.1, 0.2), group = "expl_id")
}
\arguments{
\item{group}{}
}
\description{
Get validation set
}
