% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_outlier.R
\name{GH_filter}
\alias{GH_filter}
\title{Filter sample out using a standardized Mahalanobis distance threshold}
\usage{
GH_filter(
  X,
  Y,
  variables,
  threshold = ifelse(missing(Y), 5, 3),
  ncp = 5,
  variance,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{X}{A data.frame of samples.}

\item{Y}{Another data.frame to which the subsetting is performed.
If `Y` is missing, the subsetting is made on `X`.}

\item{variables}{A vector of selection of variables on which PCA is performed.
Not mandatory. Could be logical(same size as the number of variables in the
data.frame) or indexes.}

\item{threshold}{A numeric. Distance below which the sample is accepted.}

\item{ncp}{An integer. Number of components to keep while performing PCA.}

\item{variance}{A numeric in [0, 1]
Total cumulative percentage of variance to be explained by
the new latent variables. Not mandatory.}

\item{verbose}{If TRUE, the function would talk to you.}

\item{...}{Not currently used.}
}
\value{
A vector of logical values (GH < threshold).
}
\description{
The function uses \code{\link[FactoMineR]{PCA}} to compute the principal
components. Then the Global H distance is computed from the projection
of the sample to the new coordinates. Each sample farther from the `threshold`
value from the barycenter will be noticed.
A vector of logical values with the same length as the number of rows of
the data.frame provided would be returned. FALSE value in the output
vector points out the corresponding records as beyond the threshold.
}
\details{
If the argument `variance` is missing, the function uses the
first `ncp` principal components to compute the GH distance. Otherwise,
if the `variance` is provided, the function keeps as many components needed
to explain the value passed through the argument `variance` of the cumulated
variance of the initial variables.

If `variable` argument is missing, no subsetting is performed.
}
\seealso{
\code{\link[FactoMineR]{PCA}}

\code{\link[stats]{mahalanobis}}
}
