% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliars.R
\name{GH_outlier}
\alias{GH_outlier}
\title{Remove row outliers based on the standardized Mahalanobis distance of
data points.}
\usage{
GH_outlier(data, variables = dplyr::everything(), std_err = 3,
  remove = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{data}{Data.frame. Candidate data for outliers detection.}

\item{std_err}{Threshold above which data is considered as outlier.}

\item{remove}{Boolean. Default TRUE. If TRUE, then the function returns a new dataset minus the assessed outliers. Otherwise, it returns a logical atomic vector specifying -  by the boolean value False - which data is an outlier.}

\item{verbose}{Boolean. Default TRUE.  If TRUE, some of the intermediate results are output to the console.}

\item{...}{Argument to pass through caret::preProcess function. e.g.
thresh. (see caret::preProcess)}
}
\description{
Remove row outliers based on the standardized Mahalanobis distance of
data points.
}
