% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDIofICDF.R
\name{HDIofICDF}
\alias{HDIofICDF}
\title{Compute Highest Density Interval limits}
\usage{
HDIofICDF(ICDFname, credMass = 0.95, tol = 1e-08, ...)
}
\arguments{
\item{ICDFname}{R’s name for the inverse cumulative density function of the distribution}

\item{credMass}{the desired mass of the HDI region.}

\item{tol}{tolerance passed to R’s optimize function}
}
\value{
HDI
}
\description{
Compute Highest Density interval limits.
}
\details{
Code from J.K.Kruschke - Doing bayesian data analysis.  Adapted and corrected from Greg Snow’s TeachingDemos package.
}
\examples{
#ICDFname must be explicitly named
HDIofICDF( qbeta , shape1 = 30 , shape2 = 12 )

#does not work
HDIofICDF( qbeta , 30 , 12 )

}
\author{
JuG
}
