% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.01.r
\name{is.01}
\alias{is.01}
\title{Check if a variable is 0/1 encoded}
\usage{
is.01(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
boolean
}
\description{
Check if a variable is 0/1 encoded
}
\examples{
testDf <- data.frame(a = runif(10),
                    b = rbinom(10, size = 1, prob = .2),
                    c = as.Date(Sys.time())- 10:1,
                    d = factor(x = rep(3,10)),
                    e = NA,
                    f = rep(0,10),
                    g = rep(1,10),
                    h = sample(x = c(0,1,NA), size = 10, replace = T),
                    i = sample(x = c(0,NA), size = 10, replace = T),
                    j = sample(x = c(1,NA), size = 10, replace = T))
sapply(testDf, FUN=is.01)
}
\author{
JuG
}
