% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDate.r
\name{isDate}
\alias{isDate}
\title{Test if a vector can be a class Date}
\usage{
isDate(x, date.format)
}
\arguments{
\item{x}{a vector}

\item{date.format}{a character string as in as.Date}
}
\value{

}
\description{
Test if a vector can be a class Date
}
\examples{
dates <- c("02/27/92", NA, "01/14/92", "02/28/92", "02/01/92")
isDate(dates, date.format= "\%m/\%d/\%y")

data <- data.frame( Date=c("10/11/2012","10/12/2012"),
                    AE=c(1211,100),
                    Percent=c(0.03,0.43),
                    Date2=c("10/27/2012","10/12/12"))
sapply(data,isDate)
sapply(data,isDate, date.format= "\%d/\%m/\%Y")
}
\author{
JuG
}
