% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{log_message}
\alias{log_message}
\title{Utility function to log formated messages into the console.}
\usage{
log_message(message, title = T)
}
\arguments{
\item{message}{A character. Message to be printed.}

\item{title}{logical. If TRUE, a nice hand-crafted frame is painted around the
undoubtedly useful message.}
}
\value{

}
\description{
Utility function to log formated messages into the console.
}
