% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.all.R
\name{merge.all}
\alias{merge.all}
\title{Merge several data.frames}
\usage{
\method{merge}{all}(by, ...)
}
\arguments{
\item{by}{variable to merge on}
}
\value{
merged data.frame
}
\description{
Merge several data.frames
}
\examples{
 a <- data.frame("USUBJID" = 1:10, val1 = rnorm(10,1,.2))
 b <- data.frame("USUBJID" = 1:10, val2 = LETTERS[1:10])
 c <- data.frame("USUBJID" = 1:10, val3 = gl(n = 5,k = 2,length = 10))
 merged<-merge.all(by="USUBJID",a,b,c)
}
\author{
JuG
}
