% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdirPaper.r
\name{mkdirPaper}
\alias{mkdirPaper}
\title{Create an organised folder for publications data}
\usage{
mkdirPaper(path, name)
}
\arguments{
\item{path}{path to folder where to create mkdirPaper}

\item{name}{name of the folder}
}
\description{
Create an organised folder for publications data
}
\examples{
path = getwd()
name = "testFolder"
mkdirPaper(path,name = name)
}
\author{
JuG
}
