% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTabOne.r
\name{myTabOne}
\alias{myTabOne}
\title{Create TableOne with NA counts}
\usage{
myTabOne(dat = dat, strata = "Var.Grp")
}
\value{

}
\description{
basis from package tableone
}
\examples{
dat <- data.frame(Var.Gr = gl(n = 2, k = 10),
   Var.A = rnorm(20),
   Var.B = factor(x = rbinom(n = 20, size = 1, .2)),
   Var.C = runif(n = 20,1,20))
dat$Var.A[sample(size = 4, 1:20)]<- NA
dat$Var.B[sample(size = 3, 1:20)]<- NA
pp2 <- myTabOne(dat = dat, strata = "Var.Gr")

myTabOne(dat=lung[,-2:-1], strata = "status")
}
\author{
JuG
}
