% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.count.r
\name{na.count}
\alias{na.count}
\title{Count missing data}
\usage{
na.count(data)
}
\arguments{
\item{data}{a vector (numeric, factor or ordered), a matrix, a data.frame or a list}
}
\value{
vector of missing value per variable
}
\description{
Count missing data
}
\details{
key function  \code{nafun <- function(x){sum(is.na(x))}}
}
\examples{
#with data.frame
data3 <- data.frame(a = c(1,2,3), b= c("e",NA,3), c = 1:3)
na.count(data3)
#with list
data4 <- list(a = c(1,2,3), b= c("e",NA))
na.count(data4)

}
\author{
JuG
}
