% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFun.R
\name{newFun}
\alias{newFun}
\title{Print a skeleton for a new function in the console}
\usage{
newFun(name, createFile = FALSE, path = NULL)
}
\arguments{
\item{name}{A name for the function}
}
\value{
void
}
\description{
Print a skeletton for a new function in the console
}
\details{
>Sys.setenv(TZ="Europe/Berlin") >Sys.setlocale("LC_TIME", "C"); to get UK date format
}
\examples{
newFun("newFunction")
newFun("newFunction",createFile = TRUE)

}
\author{
JuG
}
