% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newRepport.r
\name{newRepport}
\alias{newRepport}
\title{Create Rnw file for data analysis repport template using LateX}
\usage{
newRepport(nom = "NOM", prenom = "PRENOM", path = NULL,
  filename = NULL)
}
\arguments{
\item{nom}{Surname}

\item{prenom}{Forename}

\item{path}{path}

\item{filename}{filename}
}
\value{
Rnw file
}
\description{
Create Rnw file for data analysis repport template using LateX
}
\details{
adapted from  Michael Schaeffer template (gmrcfun package)
}
\examples{
newRepport(nom="G",prenom="J",filename = NULL,path=NULL)
newRepport(nom="G",prenom="J",path=getwd(),filename= "test")
}
\author{
JuG
}
