% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newRmdRepport.R
\name{newRmdRepport}
\alias{newRmdRepport}
\title{Create Rmd file for data analysis repport template using Rmd LateX pdf}
\usage{
newRmdRepport(nom = "NOM", prenom = "PRENOM", path = NULL, filename = NULL)
}
\arguments{
\item{nom}{Surname}

\item{prenom}{Forename}

\item{path}{path}

\item{filename}{filename}
}
\value{
Rnw file
}
\description{
Create Rmd file for data analysis repport template using Rmd/LateX-pdf
}
\details{
adapted from  Michael Schaeffer template (gmrcfun package)
}
\examples{
newRmdRepport(nom="G",prenom="J",filename = NULL,path=NULL)
newRmdRepport(nom="G",prenom="J",path=getwd(),filename= "test")
}
\author{
JuG
}
