% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newTemplate.r
\name{newTemplate}
\alias{newTemplate}
\title{Create a R script template}
\usage{
newTemplate(path, name, createFile = F)
}
\arguments{
\item{path}{R script path}

\item{name}{R script name}

\item{createFile}{boolean}
}
\description{
Create a R script template
}
\examples{
newTemplate(name="test")

}
\author{
JuG
}
