% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsB.R
\name{nlsB}
\alias{nlsB}
\title{Determine the nonlinear (weighted) MCMC estimates of the parameters of a nonlinear model.}
\usage{
nlsB(formula, data, graphOutput = TRUE, nIter = 10000, thin = 1)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which blm is called.}

\item{graphOutput}{regression parameters graphical output (MCMC Trace and posterior density)}

\item{nIter}{number of iterations}

\item{thin}{thinning interval for monitors}
}
\value{
non-linear regression parameters
}
\description{
Determine the nonlinear (weighted) MCMC estimates of the parameters of a nonlinear model.
}
\details{
Models for lm are specified symbolically. A typical model has the form response ~ terms where response is the (numeric) response vector and terms is a series of terms which specifies a linear predictor for response.
}
\examples{
data(mtcars)
summary(lm(mpg~ cyl + vs+gear+carb,data=mtcars))
lmB(mpg~ cyl + vs+gear+carb,data=mtcars,nIter=50000)
lmB(mpg~ .,data=mtcars,nIter=50000,graphOutput=FALSE)
}
\author{
JuG
}
