% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openRefine.r
\name{openRefine}
\alias{openRefine}
\title{Connect to openRefine API}
\usage{
openRefine(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
void
}
\description{
Connect to openRefine API to import, export or delete a project in OpenRefine directly from an R script
}
\details{
Connect to openRefine API \url{https://cran.r-project.org/web/packages/rrefine/vignettes/rrefine-vignette.html}. openRefine must be installed on the local machine \url{http://openrefine.org/download.html}.
}
\examples{
#devtools::install_github("vpnagraj/rrefine")
library(rrefine)
lateformeeting[1:10,]
openRefine(lateformeeting)

}
\author{
JuG
}
