% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliars.R
\name{outliars}
\alias{outliars}
\title{Outlier detection for quantitative data}
\usage{
outliars(data, std_err = 3, remove = TRUE, verbose = TRUE)
}
\arguments{
\item{data}{Data.frame. Candidate data for outliers detection.}

\item{std_err}{Integer. Which specifies how far from the mean a sample should be for being considered as an outlier.}

\item{remove}{Boolean. Default TRUE. If TRUE, then the function returns a new dataset minus the assessed outliers. Otherwise, it returns a logical atomic vector specifying -  by the boolean value False - which data is an outlier.}

\item{verbose}{Boolean. Default TRUE.  If TRUE, some of the intermediate results are output to the console.}
}
\value{
See remove argument.
}
\description{
Outlier detection for quantitative data
}
