% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piEst.R
\name{piEst}
\alias{piEst}
\title{Estimate population proportion and interval limits}
\usage{
piEst(
  y = 12,
  n = 20,
  beta.priors = c(0.5, 0.5),
  credMass = 0.95,
  showPlot = FALSE,
  lateX = FALSE,
  ...
)
}
\arguments{
\item{y}{counts of successes}

\item{n}{counts of trials}

\item{beta.priors}{vector of 2 non-negative parameters of the Beta prior distribution (shapes - by default Jeffrey's c(0.5,0.5))}

\item{credMass}{mass of the HDI region}

\item{showPlot}{if TRUE, show a graphical representation of Pi distribution}

\item{lateX}{if TRUE, print lateX output}
}
\value{
Proportion estimates with HDI, CI (quantile c(.025,0.975)), and binomial CI (from binom.test)
}
\description{
Estimate population proportion and interval limits from observede counts of succes and trials.
}
\examples{
piEst(y = 3, n = 12, showPlot = TRUE)
piEst(y = 3, n = 12, lateX = F)

}
\author{
JuG
}
