% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorMat.R
\name{plotCorMat}
\alias{plotCorMat}
\title{Plot correlation matrix}
\usage{
plotCorMat(cormat, cexlab = 12, angle.label = 45)
}
\arguments{
\item{cormat}{correlation matrix}

\item{cexlab}{label size}

\item{angle.label}{cex.label size}
}
\value{
plot
}
\description{
Plot correlation matrix
}
\examples{

 #correlation matrix - simulated (or alternative fact)
 corMat <- matrix(runif(400,-1,1),ncol=20)
 diag(corMat)<-1
 colnames(corMat) <- rownames(corMat) <- paste("Vble",1:20,sep='')
 #plot CorMat
 plotCorMat(corMat)

 # Correlation matrix from mtcars
 corMat2 <-cor(mtcars)
 plotCorMat(corMat2)
}
\author{
JuG
}
