% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlm.r
\name{plotlm}
\alias{plotlm}
\title{Plot 2D regression line with confidence intervals}
\usage{
plotlm(x, y, data = NULL, prediction = FALSE, ...)
}
\arguments{
\item{prediction}{if TRUE, prediction intervals are drawn}
}
\description{
Plot 2D regression line with confidence intervals
}
\examples{
nbPoints = 20
x <-seq(1,10, length = nbPoints)
y <- 3 * x + rnorm(nbPoints, 0,6) + 10
plotlm(x,y,ylim=c(0,50),main="plotlm example",prediction=T)
}
\author{
JuG
}
