% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlmB.r
\name{plotlmB}
\alias{plotlmB}
\title{Plot 2D regression line from lmB}
\usage{
plotlmB(x, y, mcmcList, data = NULL, ...)
}
\arguments{
\item{mcmcList}{a mcmc list (out put from lmB)}

\item{prediction}{if TRUE, prediction intervals are drawn}
}
\description{
#' Plot 2D regression line from lmB
}
\examples{
nbPoints = 20
x <-seq(1,10, length = nbPoints)
y <- 3 * x + rnorm(nbPoints, 0,6) + 10
outlmB <- lmB(formula = y~x,returnCodaSamples = T)
plotlmB(x,y,mcmcList = outlmB, ylim=c(0,50))
}
\author{
JuG
}
