% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurv.r
\name{plotSurv}
\alias{plotSurv}
\title{Plot survival curves}
\usage{
plotSurv(fit, pval = TRUE, ...)
}
\arguments{
\item{fit}{a survfit object}
}
\value{
graph
}
\description{
Plot survival curves
}
\details{
code from \url{http://www.sthda.com/english/wiki/survival-analysis-basics}
}
\examples{
require(survival)
fit <- survfit(Surv(time, status) ~ sex, data = lung)
plotSurv(fit)
fit <- survfit(Surv(time, status) ~ 1, data = lung)
}
\author{
JuG
}
