% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relWeights.R
\name{relWeights}
\alias{relWeights}
\title{Functions for predictor selection}
\usage{
relWeights(fit, ...)
}
\arguments{
\item{fit}{a models lm object.}
}
\value{
graphic for rredictor selection
}
\description{
Predictor selection.The total amount of variance accounted for by the model is divided amoong the predicator variables.
}
\examples{

form = as.formula("Fertility ~ .")
allsubreg(formula = form, data = swiss, nbest=6)

if(!require('MASS')){install.packages('MASS')}
library('MASS')
fit1 <- lm(Fertility ~ ., data = swiss)
relWeights(fit1, col="lightgrey",las=1)


}
\author{
JuG
}
