% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subReg.R
\name{subReg}
\alias{subReg}
\title{Functions for model selection}
\usage{
subReg(formula, data, nbest)
}
\arguments{
\item{formula}{The models fit by, e.g., the lm and glm functions  specified in a compact symbolic form.}

\item{data}{data.frame}

\item{nbest}{number of subsets of each size to record}
}
\value{
graphic for best model selection based on Adjusted R-squre
}
\description{
Model selection by exhaustive search. All subset regression are evaluated to help choosing the best minimalist model.    R in action p 211.
}
\examples{

form = as.formula("Fertility ~ .")

if(!require('MASS')){install.packages('MASS')}
library('MASS')
fit1 <- lm(form, data = swiss)
stepAIC(fit1,direction = "both")

subReg(form, data=swiss, nbest = 3)

relweights(fit1, col="lightgrey",las=1)

}
\author{
JuG
}
