% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryOR.R
\name{summaryOR}
\alias{summaryOR}
\title{Odds ratio summary statistics
#' @author JuG}
\usage{
summaryOR(glm.fit, lateX = TRUE, graph = FALSE, digits = 3, ...)
}
\arguments{
\item{glm.fit}{an object of class "glm.fit" calculated with a "binomial "link function.}

\item{lateX}{if TRUE, return lateX xtable of the result table}

\item{graph}{if TRUE, plot the coefficients of a model with \emph{broom} and \emph{ggplot2}}
}
\description{
Odds ratio summary statistics
}
\details{
Provide Odds ratio  (exponential of the glm.fit coefficients) and their confidence intervals (given by \code{\link{confint}}). Stars are "***" if p-value < 0.001, "**" if 0.001 < p-value <= 0.01, "*" if 0.01 < p-value <= 0.05, "." if 0.05 < p-value <=0.1
}
\examples{
y <- rbinom(n = 50, size = 1, prob = .3)
x1 <- rnorm(50, 5,1)
x2 <- runif(50, 0,5) * (1 + 2*y)
x3 <- rnorm(50, 12,5)
 Data <- data.frame(y,x1,x2,x3)

 logist <- glm(y~., data=Data, family="binomial")
 summary(logist)
 summaryOR(glm.fit=logist,lateX=FALSE,graph=FALSE, digits=4)
}
