% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_outlier_test.R
\name{t_outlier_test}
\alias{t_outlier_test}
\alias{t_outlier_test.formula}
\alias{t_outlier_test.recipe}
\alias{t_outlier_test.default}
\title{Run a t-outlier test (S3)}
\usage{
t_outlier_test(x, data, ...)

\method{t_outlier_test}{default}(...)

\method{t_outlier_test}{formula}(
  x,
  data,
  method,
  group = NULL,
  k = NULL,
  predictors = dplyr::everything(),
  std_err = 3,
  remove = TRUE,
  verbose = TRUE,
  ...
)

\method{t_outlier_test}{recipe}(
  x,
  data,
  method,
  group,
  k,
  predictors = dplyr::everything(),
  std_err = 3,
  remove = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Either a recipe or a formula. Depending on x type, the appropriate
fuction is called.}

\item{data}{Data.frame. Candidate data for outliers detection.}

\item{...}{Additional arguments for caret::train.}

\item{method}{A list of lists. Each sublists contain the method name and some related arguments.}

\item{std_err}{Integer. Which specifies how far from the mean a sample should be for being considered as an outlier.}

\item{remove}{Boolean. Default TRUE. If TRUE, then the function returns a new dataset minus the assessed outliers. Otherwise, it returns a logical atomic vector specifying -  by the boolean value False - which data is an outlier.}

\item{verbose}{Boolean. Default TRUE.  If TRUE, some of the intermediate results are output to the console.}
}
\value{
See remove argument.
}
\description{
The purpose of a t-outlier test is to detect potential outliers from the analysis of residuals based on a first predictive analysis.
}
\seealso{
\code{\link[caret]{train}}, \code{\link[recipes]{recipe}}
}
\author{
Anthony Tedde
}
