% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliars.R
\name{t_test_outlier}
\alias{t_test_outlier}
\title{Title}
\usage{
t_test_outlier(formula, data, method, verbose = T, ...)
}
\arguments{
\item{formula}{depedent var ~ independent ones.}

\item{data}{Candidate data.}

\item{method}{List of methods to apply.}

\item{...}{}
}
\value{
a logical atomic vector of row outliers based on the combination of
all performed t-test outliers detection.
}
\description{
Title
}
