% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempTab.r
\name{tempTab}
\alias{tempTab}
\title{Draft code script for table one}
\usage{
tempTab(gv = gv, strata = NULL)
}
\arguments{
\item{gv}{data.frame name}

\item{strata}{strata (optional)}
}
\value{

}
\description{
Print draft code script for table one
}
\examples{
gv <- "data.frame.Name"
tempTab(gv=gv)
}
\author{
JuG
}
