% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unaccent.r
\name{unaccent}
\alias{unaccent}
\title{Transform accented into unaccented characters}
\usage{
unaccent(text)
}
\arguments{
\item{text}{an in put that can be coerced as character}
}
\value{
text
}
\description{
Transform accented into unaccented characters
}
\details{
from \url{http://data.hypotheses.org/564}
}
\examples{
unaccent(c("é","à","è","ù","ç",'ü',"a,"))
unaccent("À Süßen (Ba.-Wü.), j'ai mangé des Curry Würstchen")
}
\author{
JuG
}
