% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Unscaled scaled data according to the former ones.}
\usage{
unscale(scaled_data, template)
}
\arguments{
\item{scaled_data}{The data to by unscaled.}

\item{template}{The template data for unscaling.}
}
\value{
A data.frame of unscaled data.
}
\description{
Unscaled scaled data according to the former ones.
}
