% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout.R
\name{workout}
\alias{workout}
\title{Workout}
\usage{
workout(args, seed = 1010, subset = T, features = T)
}
\arguments{
\item{args}{All the argumnets to pass through the caret::train function}

\item{seed}{A seed. For reproducibility.}

\item{subset}{The row subset of data to use for the analysis. All by default.}

\item{features}{A subset of regressors to use. All by default.}
}
\value{
A list of caret object.
}
\description{
A function to perform models training accross with different
data. Simultaneously.
}
