% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_validation_set.R
\name{get_validation_set}
\alias{get_validation_set}
\title{Get validation set}
\usage{
get_validation_set(
  data,
  variables = dplyr::everything(),
  bound = c(0.1, 0.2),
  group
)
}
\arguments{
\item{data}{Data.frame. Candidate data.frame.}

\item{variables}{Character. What variables to consider in PCA.}

\item{bound}{Integer. Lower and upper boundary of data proportion to
be covered by the validation set.}

\item{group}{Character. Pivot variable which discriminate each group.}

\item{predicate}{Function. Function to use for subset selection}
}
\value{

}
\description{
Creation of validation sets based on some criterions, such as the proportion
of data going to the validation set `bound` ans
}
